dbt_query_tool_agent/requirements.txt
google-adk
google-cloud-storage
google-generativeai
Pillow
dbt-bigquery
fastapi
dbt-core
pydantic==2.10.6